refactor(hero): integration hero into home page component
